new fix

+++ yatube/posts/views.py
@@ -86,7 +86,9 @@ def profile(request, username):
     page = pagination(request, user_posts)
 
     follow_flag = False
-    if Follow.objects.filter(user=request.user, author=profile_user).exists():
+    if ((request.user.is_authenticated) and
+        Follow.objects.filter(user=request.user, author=profile_user).exists()    
+    ):
         follow_flag = True
 
     return render(request, 'posts/profile.html',
@@ -96,10 +98,11 @@ def profile(request, username):
 
 def post_view(request, username, post_id):
     post = get_object_or_404(Post, author__username=username, id=post_id)
-    if request.user.is_authenticated:
-        form = CommentForm(request.POST or None)
+    # if request.user.is_authenticated:
+    form = CommentForm(request.POST or None)
     return render(request, 'posts/post.html',
                   {'post': post,
+                   'author': post.author.username,
                    'form': form,
                    'comments': post.comments.all()})
 
@@ -131,12 +134,12 @@ def add_comment(request, username, post_id):
         new_comment.author = request.user
         new_comment.post = post
         new_comment.save()
-        return redirect('post', username=post.author.username,
-                        post_id=post.id)
-    return render(request, 'posts/post.html',
+    return redirect('post', username=post.author.username,
+                    post_id=post.id)
+    """return render(request, 'posts/post.html',
                   {'post': post,
                    'form': form,
-                   'comments': post.comments.all()})
+                   'comments': post.comments.all()})"""
 
 
 @login_required
@@ -179,5 +182,6 @@ def profile_follow(request, username):
 @login_required
 def profile_unfollow(request, username):
     profile_user = get_object_or_404(User, username=username)
-    Follow.objects.get(user=request.user, author=profile_user).delete()
+    if Follow.objects.filter(user=request.user).filter(author=profile_user).exists():
+        Follow.objects.get(user=request.user, author=profile_user).delete()
     return redirect('profile', username=username)
